# api/swagger/v1/todo-service.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "ToDo service",
    "version": "1.0",
    "contact": {
      "name": "go-grpc-http-rest-microservice",
      "url": "https://github.com/sanjid133/go-grpc-http-rest-microservice",
      "email": "sanjid133@gmail.com"
    }
  },
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/v1/todo/all": {
      "get": {
        "operationId": "ReadAll",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ReadAllResponse"
            }
          },
          "404": {
            "description": "Returned when values not found",
            "schema": {
              "format": "string"
            }
          }
        },
        "parameters": [
          {
            "name": "api",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "ToDoService"
        ]
      }
    },
...
  },
  "definitions": {
...
    "v1ReadAllResponse": {
      "type": "object",
      "properties": {
        "api": {
          "type": "string"
        },
        "toDos": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/v1ToDo"
          }
        }
      }
    },
...
    "v1ToDo": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        },
        "title": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "reminder": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
...
  }
}